remove unused pages and fix navigation for /assistants
